move wer helpers to api layer
